\name{LMMCMC-package}
\alias{LMMCMC-package}
\docType{package}
\title{LMMCMC: MCMC for Gaussian linear models.}
\description{
Currently devoted to just estimating model posterior probabilities. Designed to be fast and have small memory overhead.
}
\seealso{
\code{\link{lm.mcmc}()}, \code{\link{model.prior-class}}, \code{\link{coef.prior-class}}.
}
\author{
Andrew Womack, \cr Maintainer: Andrew Womack <womack.andrew@gmail.com>
}
\concept{mcmc methods}
\keyword{package}
\keyword{regression}
