\name{LMMCMC}
\alias{LMMCMC}
\title{LMMCMC: MCMC for Gaussian linear models.}
\description{
Currently devoted to just estimating model posterior probabilities. Designed to be fast and have small memory overhead.
}
\details{
_PACKAGE
}
\seealso{
\code{\link{lm.mcmc}()}, \code{\link{md.prior}()}, \code{\link{custom.prior}()}
}
\author{
Andrew Womack, \cr Maintainer: Andrew Womack <womack.andrew@gmail.com>
}
\concept{mcmc methods}
\keyword{package}
\keyword{regression}
